Uploaded Initial Two Files
This is the initial commit and contains the two pieces of code I am currently using to display two different types of graphs. The first graph (All News Sources Graphed) is a folder with an index.html and it draws multiple circles at different y-positions to represent the popularity, and places bubbles at various x-positions to represent bias.  The second piece of code (Categorized Bubble) is much more simple and is basically already fully completed. Still have work to do for the first graph.